% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoplen.R
\name{nanoplen}
\alias{nanoplen}
\title{Run NanopLen differential length on a file of lengths per transcript}
\usage{
nanoplen(
  data_file,
  metadata,
  test = "t",
  baseline = "Control",
  logscale = FALSE,
  params = NULL,
  norm = FALSE
)
}
\description{
NanopLen tests for significant differences in lengths, e.g. transcript length
or Poly(A) tail length, between two categories.
}
\section{Expected inputs}{

\describe{
  \item {data file}{A file of lengths with these columns in order: library id, gene/transcript id, length}
  \item {metadata}{Metadata with these columns in order: library id, condition, and any optional additional columns}
  \item {test}{Type of statistical test: 't' for t-test, 'm' for linear mixed model, 'w' for Wilcoxon}
  \item {baseline}{String that indicates the baseline variable for the 'condition' variable in the metadata}
  \item {logscale}{ indicates if length should be on log2scale}
  \item {params}{Extra parameters to use in the model, done in model notation (e.g. to add age and sex and their interaction to the model, the input is 'age+sex+age*sex') }
}
}

\section{Expected output}{

A results file with the following columns:
\describe{
  \item {id}{corresponds to gene/transcript as given in data file}
  \item {meandiff}{the mean difference in length between conditions, shown if logscale = F}
  \item {log2fc}{log2 fold change in length between conditions, shown if logscale = T}
  \item {pvalue}{p-values for significance of differential length test}
  \item {qvalue}{Adjusted p-values using the Bonferroni-Hochberg method}
  \item {n.baseline}{Number of observations for baseline condition. "baseline" will be replaced with user-defined baseline string}
  \item {n.alt} {Number of observations for alternate condition}
  \item {mean_length.baseline}{Mean length of reads in baseline condition. "baseline" will be replaced with user-defined baseline string}
  \item {mean_length.baseline}{Mean length of reads in alternate condition.}
}
}

\section{Tests}{

\describe{
  \item {t}{The most basic t-test between conditions. Supports additional parameters which makes this functionally equivalent to
simple linear regression}
  \item {m}{Linear Mixed Model, testing for difference between conditions with a random effect on the libraries. Supports additional parameters.}
  \item {w}{Wilcoxon Rank-Sum test, testing for difference between conditions. Does not support extra parameters.}
}
}

